requirements.txt
fastapi
pydantic
pydantic-settings
weaviate-client
python-jose
langchain
uvicorn
aiofiles